requirements.txt
ruamel.yaml
jsonextended>=0.5.4
jsonschema
paramiko